Add ga_debug.js option
This adds a debug bool param to trigger the Google Analytics ga_debug.js as explained https://developers.google.com/analytics/resources/articles/gaTrackingTroubleshooting#gaDebug

If accepted, it should also be documented in the wiki under advanced usage. Mostly just a note on enabling it and the GA link so folk know what it can do.